Refactored the "Directory without permissions test so that it creates a temporary one

+++ net/src/mbedtls/test/data_loader.c
@@ -19,6 +19,8 @@
 
 #include <unistd.h>
 
+#include <sys/stat.h>
+
 AVS_UNIT_TEST(backend_mbedtls, chain_loading_from_file) {
     mbedtls_x509_crt *chain = NULL;
 
@@ -49,19 +51,33 @@ AVS_UNIT_TEST(backend_mbedtls, chain_loading_from_path) {
     mbedtls_x509_crt_free(chain);
 
     // Empty directory case.
-    char name[] = "/tmp/empty-XXXXXX";
-    (void) mkdtemp(name);
-    const avs_net_trusted_cert_info_t empty_dir =
-            avs_net_trusted_cert_info_from_path(name);
-    int retval = _avs_net_mbedtls_load_ca_certs(&chain, &empty_dir);
-    (void) rmdir(name);
-    AVS_UNIT_ASSERT_SUCCESS(retval);
+    {
+        char name[] = "/tmp/empty-XXXXXX";
+        AVS_UNIT_ASSERT_NOT_NULL(mkdtemp(name));
+        const avs_net_trusted_cert_info_t empty_dir =
+                avs_net_trusted_cert_info_from_path(name);
+        int retval = _avs_net_mbedtls_load_ca_certs(&chain, &empty_dir);
+        (void) rmdir(name);
+        AVS_UNIT_ASSERT_SUCCESS(retval);
+    }
 
     // Directory without permissions - hopefully nobody runs tests as root.
-    const avs_net_trusted_cert_info_t no_permissions_dir =
-            avs_net_trusted_cert_info_from_path("/root");
-    AVS_UNIT_ASSERT_FAILED(
-            _avs_net_mbedtls_load_ca_certs(&chain, &no_permissions_dir));
+    {
+        char name[] = "/tmp/locked-XXXXXX";
+        AVS_UNIT_ASSERT_NOT_NULL(mkdtemp(name));
+        int retval = chmod(name, 0);
+        if (retval) {
+            (void) rmdir(name);
+        }
+        AVS_UNIT_ASSERT_SUCCESS(retval);
+        const avs_net_trusted_cert_info_t no_permissions_dir =
+                avs_net_trusted_cert_info_from_path(name);
+        retval = _avs_net_mbedtls_load_ca_certs(&chain,
+                                                &no_permissions_dir);
+        (void) rmdir(name);
+        AVS_UNIT_ASSERT_FAILED(retval);
+    }
+
     free(chain);
 }
